blocking and non blocking code
Blocking refers to operations that block further execution until that operation finishes while non-blocking relates to code that doesn't stop execution. Or as Node. js docs put it, blocking is when the implementation of additional JavaScript in the Node. js process must wait until a non-JavaScript operation completes. Blocking refers to operations that block further execution until that operation finishes while non-blocking relates to code that doesn't stop execution. Or as Node. js docs put it, blocking is when the implementation of additional JavaScript in the Node. js process must wait until a non-JavaScript operation completes.